Fix mutant with default value
