+++ tests/Base/ManagerTransactionTest.php
@@ -7,40 +7,14 @@ namespace Yiisoft\Rbac\Cycle\Tests\Base;
 use RuntimeException;
 use Yiisoft\Rbac\AssignmentsStorageInterface;
 use Yiisoft\Rbac\Cycle\ItemsStorage;
-use Yiisoft\Rbac\Cycle\Manager;
 use Yiisoft\Rbac\ItemsStorageInterface;
-use Yiisoft\Rbac\RuleFactoryInterface;
-use Yiisoft\Rbac\Tests\Common\ManagerTestConfigurationTrait;
 use Yiisoft\Rbac\Tests\Support\FakeAssignmentsStorage;
-use Yiisoft\Rbac\Tests\Support\SimpleRuleFactory;
 
-abstract class ManagerTransactionTest extends TestCase
+abstract class ManagerTransactionTest extends ManagerTest
 {
-    use ManagerTestConfigurationTrait;
-
     protected function setUp(): void
     {
         $this->createSchemaManager()->ensureTables();
-    }
-
-    protected function populateDatabase(): void
-    {
-        // Skip
-    }
-
-    protected function createManager(
-        ?ItemsStorageInterface $itemsStorage = null,
-        ?AssignmentsStorageInterface $assignmentsStorage = null,
-        ?RuleFactoryInterface $ruleFactory = null,
-        bool $enableDirectPermissions = false
-    ): Manager {
-        return new Manager(
-            $itemsStorage ?? $this->createItemsStorage(),
-            $assignmentsStorage ?? $this->createAssignmentsStorage(),
-            $ruleFactory ?? new SimpleRuleFactory(),
-            $this->getDatabase(),
-            $enableDirectPermissions,
-        );
     }
 
     protected function createItemsStorage(): ItemsStorageInterface
